Display and prompt when multiple files match
Multiple matches from the 'find' command resulted in mpv trying to play
them all and, as a result, failing. Now, more than one match results in
all options being shown and selected from. Resolves #4 and, as no 'find'
matches now makes the script exit with a message to that effect,
resolves #3 as well.