Add sleep to destroy for dwmblocks timing
dwmblocks updating done via the kill command was happening too soon
after killing the daemon meaning the bar remained with a blank space for
the torrents until next update. Adding the sleep is an inelegant fix for
this issue but a fix nonetheless. Also added the dwmblocks-based kill
command to the path where no torrents are present.